requirements.txt
# LangChain core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.2.0


# Environment variable management
python-dotenv>=1.0.0

